Splitted errors and auxiliar functions from def.rs to error.rs and util.rs. Updated use of crates in every files
